requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
dotenv==0.9.9
einops==0.8.1
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.112.0
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
huggingface-hub==0.24.5
idna==3.4
image_gen_aux @ git+https://github.com/asomoza/image_gen_aux.git@c51646c3e068b06ba99f113ac768e64212f0981e
imageio==2.34.2
injector==0.22.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
langdetect==1.0.9
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.3
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==2.3.0
nepali-datetime==1.0.8.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
polars==1.33.0
prompt_toolkit==3.0.47
propcache==0.3.1
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.2.1
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
pytorch-lightning==2.5.1
pytz==2025.2
PyWavelets==1.9.0
pywin32==310
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
rpds-py==0.27.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
shapely==2.0.5
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spandrel==0.4.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.49.1
supervision==0.26.1
sympy==1.12
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.7.24
timm==1.0.19
tokenizers==0.19.1
toml==0.10.2
torchmetrics==1.7.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.192
ultralytics-thop==2.0.17
urllib3==1.26.13
uvicorn==0.30.5
wandb==0.19.10
watchdog==6.0.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0

